Refactor helpers into utils
